Update MSM in aipw_other_diarrhea_2
I added in the second line of the EIF when computing EIF of MSM parameters.

I also slightly modified the propensity score truncation so that propensities for the second line of the EIF are also truncated.

Finally, I fixed the same bug for scaling the EIFs by n as in the last commit.